Phase 2

+++ ESRGC.DLLR.EARN/Controllers/CommunicationController.cs
@@ -129,8 +129,10 @@ and/or view this user’s Organizational Profile for more information.",
             return false;
         })
         .ToList();
-      deleteNotes.ForEach(x => _workUnit.NotificationRepository.DeleteEntity(x));
-      _workUnit.saveChanges();
+      if (ModelState.IsValid) {
+        deleteNotes.ForEach(x => _workUnit.NotificationRepository.DeleteEntity(x));
+        _workUnit.saveChanges();
+      }
       return PartialView(notifcations);
     }
     [VerifyAccount]
@@ -151,6 +153,8 @@ and/or view this user’s Organizational Profile for more information.",
       _workUnit.saveChanges();
       return returnToUrl(Request.UrlReferrer.LocalPath, Url.Action("Detail", "Profile"));
     }
+
+    [VerifyAccount]
     public ActionResult ViewNotification(int notificationID) {
       var notification = _workUnit.NotificationRepository.GetEntityByID(notificationID);
       ActionResult result = null;
@@ -168,22 +172,26 @@ and/or view this user’s Organizational Profile for more information.",
           break;
       }
       //mark all as read
-      notification.Account.Notifications.ToList().ForEach(x => {
-        notification.IsRead = true;
-        _workUnit.NotificationRepository.UpdateEntity(x);
-      });
-      _workUnit.saveChanges();
+      if (ModelState.IsValid) {
+        notification.Account.Notifications.ToList().ForEach(x => {
+          notification.IsRead = true;
+          _workUnit.NotificationRepository.UpdateEntity(x);
+        });
+        _workUnit.saveChanges();
+      }
       if (result == null)
         return RedirectToAction("index", "home");
 
       return result;
     }
+
     [VerifyAccount]
     public ContentResult RequestCount() {
       var requests = CurrentAccount.ReceivedRequests
         .Where(x => x.Status.ToLower() == "new").ToList();
       return Content(requests.Count().ToString());
     }
+
     [VerifyAccount]
     public ActionResult Requests() {
       var requests = CurrentAccount
@@ -191,6 +199,7 @@ and/or view this user’s Organizational Profile for more information.",
         .ToList();
       return View(requests);
     }
+
     [SendNotification]
     public ActionResult AcceptRequest(int requestID) {
       var request = _workUnit.RequestRepository.GetEntityByID(requestID);
@@ -292,6 +301,7 @@ and/or view this user’s Organizational Profile for more information.",
     [HttpPost]
     [VerifyProfile]
     [VerifyProfilePartnership]
+    [ValidateAntiForgeryToken]
     [SendNotification]
     public ActionResult PostComment(int partnershipID, string comment, string returnUrl) {
       var partnership = _workUnit.PartnershipRepository.GetEntityByID(partnershipID);
@@ -329,7 +339,8 @@ and/or view this user’s Organizational Profile for more information.",
       }
     }
     [HttpPost]
-    [VerifyProfile]
+    [VerifyProfile]   
+    [ValidateAntiForgeryToken]
     public ActionResult DeleteComment(int commentID, string returnUrl) {
       try {
         var comment = _workUnit.CommentRepository.GetEntityByID(commentID);
